feature(bento-app-banner): moved component to top-level folder
