+++ css/Z_INDEX.md
@@ -61,12 +61,12 @@
 | `amp-mega-menu`                                                                                  | 1000       | [extensions/amp-mega-menu/0.1/amp-mega-menu.css](/extensions/amp-mega-menu/0.1/amp-mega-menu.css)                                                                       |
 | `amp-user-notification`                                                                          | 1000       | [extensions/amp-user-notification/0.1/amp-user-notification.css](/extensions/amp-user-notification/0.1/amp-user-notification.css)                                       |
 | `i-amphtml-app-banner-top-padding`                                                               | 15         | [extensions/amp-app-banner/0.1/amp-app-banner.css](/extensions/amp-app-banner/0.1/amp-app-banner.css)                                                                   |
-| `bannerPadding`                                                                                  | 15         | [extensions/amp-app-banner/1.0/component/component.jss.js](/extensions/amp-app-banner/1.0/component/component.jss.js)                                                   |
+| `bannerPadding`                                                                                  | 15         | [extensions/amp-app-banner/1.0/component.jss.js](/extensions/amp-app-banner/1.0/component.jss.js)                                                   |
 | `.amp-app-banner-dismiss-button`                                                                 | 14         | [extensions/amp-app-banner/0.1/amp-app-banner.css](/extensions/amp-app-banner/0.1/amp-app-banner.css)                                                                   |
-| `dismiss`                                                                                        | 14         | [extensions/amp-app-banner/1.0/component/component.jss.js](/extensions/amp-app-banner/1.0/component/component.jss.js)                                                   |
+| `dismiss`                                                                                        | 14         | [extensions/amp-app-banner/1.0/component.jss.js](/extensions/amp-app-banner/1.0/component.jss.js)                                                   |
 | `amp-app-banner`                                                                                 | 13         | [extensions/amp-app-banner/0.1/amp-app-banner.css](/extensions/amp-app-banner/0.1/amp-app-banner.css)                                                                   |
 | `amp-app-banner`                                                                                 | 13         | [extensions/amp-app-banner/1.0/amp-app-banner.css](/extensions/amp-app-banner/1.0/amp-app-banner.css)                                                                   |
-| `banner`                                                                                         | 13         | [extensions/amp-app-banner/1.0/component/component.jss.js](/extensions/amp-app-banner/1.0/component/component.jss.js)                                                   |
+| `banner`                                                                                         | 13         | [extensions/amp-app-banner/1.0/component.jss.js](/extensions/amp-app-banner/1.0/component.jss.js)                                                   |
 | `amp-sticky-ad-top-padding`                                                                      | 12         | [extensions/amp-sticky-ad/1.0/amp-sticky-ad.css](/extensions/amp-sticky-ad/1.0/amp-sticky-ad.css)                                                                       |
 | `style`                                                                                          | 11         | [extensions/amp-ad-network-doubleclick-impl/0.1/amp-ad-network-doubleclick-impl.js](/extensions/amp-ad-network-doubleclick-impl/0.1/amp-ad-network-doubleclick-impl.js) |
 | `amp-sticky-ad`                                                                                  | 11         | [extensions/amp-sticky-ad/1.0/amp-sticky-ad.css](/extensions/amp-sticky-ad/1.0/amp-sticky-ad.css)                                                                       |
