+++ extensions/amp-app-banner/1.0/base-element.js
@@ -1,7 +1,7 @@
 import {PreactBaseElement} from '#preact/base-element';
 
-import {BentoAppBanner} from './component/component';
-import {CSS as COMPONENT_CSS} from './component/component.jss';
+import {BentoAppBanner} from './component';
+import {CSS as COMPONENT_CSS} from './component.jss';
 
 export class BaseElement extends PreactBaseElement {}
