+++ extensions/amp-app-banner/1.0/component.jss.js
