+++ extensions/amp-app-banner/1.0/component.type.js
