+++ extensions/amp-app-banner/1.0/utils/android.js
